Combine RemoteNode and ProtocolHandler (#1520)
