--- tests/neo.UnitTests/Network/P2P/UT_ProtocolHandler.cs
@@ -1,42 +1,0 @@
-using Akka.TestKit.Xunit2;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Neo.Network.P2P;
-using Neo.Network.P2P.Capabilities;
-using Neo.Network.P2P.Payloads;
-
-namespace Neo.UnitTests.Network.P2P
-{
-    [TestClass]
-    public class UT_ProtocolHandler : TestKit
-    {
-        [TestCleanup]
-        public void Cleanup()
-        {
-            Shutdown();
-        }
-
-        [TestMethod]
-        public void ProtocolHandler_Test_SendVersion_TellParent()
-        {
-            var senderProbe = CreateTestProbe();
-            var parent = CreateTestProbe();
-            var protocolActor = ActorOfAsTestActorRef(() => new ProtocolHandler(TestBlockchain.TheNeoSystem), parent);
-
-            var payload = new VersionPayload()
-            {
-                UserAgent = "".PadLeft(1024, '0'),
-                Nonce = 1,
-                Magic = 2,
-                Timestamp = 5,
-                Version = 6,
-                Capabilities = new NodeCapability[]
-                {
-                    new ServerCapability(NodeCapabilityType.TcpServer, 25)
-                }
-            };
-
-            senderProbe.Send(protocolActor, Message.Create(MessageCommand.Version, payload));
-            parent.ExpectMsg<VersionPayload>();
-        }
-    }
-}
