+++ tests/neo.UnitTests/Network/P2P/UT_RemoteNode.cs
@@ -2,6 +2,7 @@ using Akka.IO;
 using Akka.TestKit.Xunit2;
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Neo.IO;
 using Neo.Network.P2P;
 using Neo.Network.P2P.Capabilities;
 using Neo.Network.P2P.Payloads;
@@ -14,8 +15,7 @@ namespace Neo.UnitTests.Network.P2P
         private static NeoSystem testBlockchain;
 
         public UT_RemoteNode()
-            : base($"remote-node-mailbox {{ mailbox-type: \"{typeof(RemoteNodeMailbox).AssemblyQualifiedName}\" }}" +
-                $"protocol-handler-mailbox {{ mailbox-type: \"{typeof(ProtocolHandlerMailbox).AssemblyQualifiedName}\" }}")
+            : base($"remote-node-mailbox {{ mailbox-type: \"{typeof(RemoteNodeMailbox).AssemblyQualifiedName}\" }}")
         {
         }
 
@@ -33,7 +33,7 @@ namespace Neo.UnitTests.Network.P2P
 
             connectionTestProbe.ExpectMsg<Tcp.Write>();
 
-            var payload = new VersionPayload()
+            var msg = Message.Create(MessageCommand.Version, new VersionPayload
             {
                 UserAgent = "".PadLeft(1024, '0'),
                 Nonce = 1,
@@ -44,10 +44,10 @@ namespace Neo.UnitTests.Network.P2P
                 {
                     new ServerCapability(NodeCapabilityType.TcpServer, 25)
                 }
-            };
+            });
 
             var testProbe = CreateTestProbe();
-            testProbe.Send(remoteNodeActor, payload);
+            testProbe.Send(remoteNodeActor, new Tcp.Received((ByteString)msg.ToArray()));
 
             connectionTestProbe.ExpectMsg<Tcp.Abort>();
         }
@@ -60,7 +60,7 @@ namespace Neo.UnitTests.Network.P2P
 
             connectionTestProbe.ExpectMsg<Tcp.Write>();
 
-            var payload = new VersionPayload()
+            var msg = Message.Create(MessageCommand.Version, new VersionPayload()
             {
                 UserAgent = "Unit Test".PadLeft(1024, '0'),
                 Nonce = 1,
@@ -71,10 +71,10 @@ namespace Neo.UnitTests.Network.P2P
                 {
                     new ServerCapability(NodeCapabilityType.TcpServer, 25)
                 }
-            };
+            });
 
             var testProbe = CreateTestProbe();
-            testProbe.Send(remoteNodeActor, payload);
+            testProbe.Send(remoteNodeActor, new Tcp.Received((ByteString)msg.ToArray()));
 
             var verackMessage = connectionTestProbe.ExpectMsg<Tcp.Write>();
